requirements.txt
annotated-types==0.7.0
anyio==4.7.0
exceptiongroup==1.2.2
fastapi==0.115.6
idna==3.10
pydantic==2.10.4
pydantic_core==2.27.2
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.34.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.20
pydantic_settings==2.7.1 
python-dotenv==1.0.1